Master groovy import, README with actual instructions
This commit adds instructions for users on how to create an L-System
model using groovy. It also changes the file structure so that one need
only import 'grooby.rb' to produce a model word. We are not yet at the
point where a word and model can be generated all at once, but that
will come soon.