Biggest problem: solved.
Interprets and draws only when needed. Solves the problems of topology,
string position, and manipulability. Set up to give nodes a substring
position so that we can update the rotation literals in the word.
That's next: in-word rotation literal updating, then next/prev
generation traversing.

Problem: Branches don't stay connected. There's a design pattern to fix
this somewhere.